+++ project/3rdparty/source/rockchip/rkmpp/mpp/hal/rkenc/h265e/hal_h265e_vepu540c.c
@@ -30,6 +30,7 @@
 
 #include "../../common/h265/hal_h265e_debug.h"
 #include "../../../common/h265e_syntax_new.h"
+#include "../../common/h265/hal_h265e_stream_amend.h"
 #include "../../common/hal_bufs.h"
 #include "../common/rkv_enc_def.h"
 #include "../common/vepu541_common.h"
@@ -1639,6 +1640,7 @@ MPP_RET hal_h265e_v540c_ret_task(void *hal, HalEncTask *task)
     hal_h265e_enter();
 
     vepu540c_h265_set_feedback(ctx, enc_task);
+    hal_h265e_amend_temporal_id(task, fb->out_strm_size);
 
     rc_info->sse = fb->sse_sum;
     rc_info->lvl64_inter_num = fb->st_lvl64_inter_num;