+++ project/3rdparty/source/rockchip/rkmpp/mpp/hal/rkenc/h265e/hal_h265e_vepu580.c
@@ -29,6 +29,7 @@
 
 #include "../../common/h265/hal_h265e_debug.h"
 #include "../../../common/h265e_syntax_new.h"
+#include "../../common/h265/hal_h265e_stream_amend.h"
 #include "../../common/hal_bufs.h"
 #include "../common/rkv_enc_def.h"
 #include "../../../codec/enc/h265/h265e_dpb.h"
@@ -3227,13 +3228,13 @@ MPP_RET hal_h265e_v580_get_task(void *hal, HalEncTask *task)
 
         frm_cfg->hal_curr_idx = ctx->syn->sp.recon_pic.slot_idx;
         frm_cfg->hal_refr_idx = ctx->syn->sp.ref_pic.slot_idx;
-
-        h265e_dpb_hal_start(ctx->dpb, frm_cfg->hal_curr_idx);
-        h265e_dpb_hal_start(ctx->dpb, frm_cfg->hal_refr_idx);
     } else {
         /* reencode path may change the frame type */
         frm_cfg = ctx->frm;
     }
+
+    h265e_dpb_hal_start(ctx->dpb, frm_cfg->hal_curr_idx);
+    h265e_dpb_hal_start(ctx->dpb, frm_cfg->hal_refr_idx);
 
     ctx->frame_type = (frm_status->is_intra) ? INTRA_FRAME : INTER_P_FRAME;
     frm_cfg->frame_type = ctx->frame_type;
@@ -3252,6 +3253,7 @@ MPP_RET hal_h265e_v580_ret_task(void *hal, HalEncTask *task)
     RK_S32 task_idx = task->flags.reg_idx;
     Vepu580H265eFrmCfg *frm = ctx->frms[task_idx];
     Vepu580H265Fbk *fb = &frm->feedback;
+    H265eSyntax_new *syn = (H265eSyntax_new *) enc_task->syntax.data;
 
     hal_h265e_enter();
 
@@ -3266,21 +3268,31 @@ MPP_RET hal_h265e_v580_ret_task(void *hal, HalEncTask *task)
                 if (i) {  //copy tile 1 stream
                     RK_U32 len = fb->out_strm_size - stream_len;
                     MppBuffer buf = frm->hw_tile_stream[i - 1];
-                    void *tile1_ptr  = mpp_buffer_get_ptr(buf);
+                    RK_U8 *tile1_ptr  = mpp_buffer_get_ptr(buf);
 
                     mpp_buffer_sync_ro_partial_begin(buf, 0, len);
+
+                    if (syn->sp.temporal_id && len > 5)
+                        tile1_ptr[5] = (tile1_ptr[5] & 0xf8) | ((syn->sp.temporal_id + 1) & 0x7);
+
                     memcpy(ptr + stream_len + offset, tile1_ptr, len);
                 } else {
                     MppBuffer buf = enc_task->output;
                     RK_U32 len = fb->out_strm_size;
+                    RK_U8 *stream_ptr = (RK_U8 *) ptr;
 
                     mpp_buffer_sync_ro_partial_begin(buf, offset, len);
+
+                    if (syn->sp.temporal_id) {
+                        stream_ptr[5] = (stream_ptr[5] & 0xf8) | ((syn->sp.temporal_id + 1) & 0x7);
+                    }
                 }
                 stream_len = fb->out_strm_size;
             }
         }
     } else {
         vepu580_h265_set_feedback(ctx, enc_task, ctx->tile_num - 1);
+        hal_h265e_amend_temporal_id(task, fb->out_strm_size);
     }
 
     rc_info->sse = fb->sse_sum;